build(engine): zero the compiler-warning inventory
A cold `linux-prod` + tests build emitted 205 warnings at 148 sites across
10 flags. It now emits none, on both the release and the debug leg, which is
the prerequisite half of #884: this makes the count zero, #884 makes warnings
fatal.

Six root causes, fixed at the cause rather than at the site:

- `-Wmissing-field-initializers` (143 emissions): five structs and one C
  struct, not 100+ brace-init sites. `ConfigEntry::unit` (53),
  `Variable::created_at` (43), `TokenError::provider_error{,_description}`
  (16), `ScenarioStep`'s four trailing fields (24) and `ExpectState::message`
  gain default member initializers - GCC does not warn for a member that
  carries one, and the values are what default construction already produced.
  `JSClassExoticMethods` is a QuickJS C struct and cannot, so it is
  value-initialized and then assigned.
- `-Wuninitialized` / `-Wmaybe-uninitialized` (8): the one genuine-UB item.
  `db::Result` and `db::MonitorSample` had bare scalars, and rows built
  field-by-field were copied with `id`, `timestamp` and `latency_ms`
  indeterminate. All are `= 0` now.
- `-Wconversion` (12, 5 sites): copy-initializing an `optional<json>` from a
  `json` puts nlohmann's `operator ValueType()` against `optional`'s
  converting constructor. `std::make_optional` direct-initializes, which
  considers the constructor alone - the answer already used in
  `utils/json.cpp`. Covers `spec_sync.cpp:376` and its four siblings in
  `openapi_export.cpp`.
- `-Wsign-conversion` (9): guarded `int` indexes into `std::array` and
  `std::vector`, plus two `char` -> `unsigned char` predicate arguments; the
  two `JS_NewObjectClass` calls drop a vestigial `static_cast<int>` that was
  converting a `JSClassID` to `int` and straight back.
- `-Wrange-loop-construct` (5): loops copying a `nlohmann::json` per
  iteration take it by reference.
- `-Wcomment` (1): prose in a block comment that opened another one.

The three false-positive families (27 emissions - `-Wdangling-reference`,
`-Wnull-dereference`, `-Warray-bounds`/`-Wstringop-overflow`) are suppressed
one function at a time through `utils/diagnostics.hpp`, which carries the
analysis, the trace and the condition for deleting each. Never widened into
`vayu_warnings`: a `-Wno-...` there would hide the next genuine instance of
the same family across the whole engine. The last family reports the same
`memcpy` under either of two flag names depending on which pass reaches it,
so both are named together.

Two clang/macOS items go with them: the two `/scripting` handlers captured a
`ctx` they never use, and `vayu_tests` linked both `GTest::gtest` and
`GTest::gtest_main`, which made Apple's linker report a duplicate library on
every macOS build.

The MSVC enumeration stays with #884, which reviews the five `/wd`
suppressions currently hiding those classes.

Closes #897

+++ engine/src/core/metrics_collector.cpp
@@ -228,7 +228,8 @@ bool MetricsCollector::claim_status_exemplar (int status_code) {
     const int slot = (status_code >= 0 && status_code < STATUS_CODE_SLOTS - 1) ?
     status_code :
     STATUS_CODE_SLOTS - 1;
-    const size_t claimed = exemplar_claims_[slot].fetch_add (1, std::memory_order_relaxed);
+    const size_t claimed =
+    exemplar_claims_[static_cast<size_t> (slot)].fetch_add (1, std::memory_order_relaxed);
     return claimed < constants::metrics_collector::EXEMPLARS_PER_STATUS;
 }
 
@@ -734,7 +735,8 @@ MetricsCollector::Percentiles MetricsCollector::sample_window_percentiles () {
 void MetricsCollector::record_status_code (int status_code) {
     if (status_code >= 0 && status_code < STATUS_CODE_SLOTS) {
         // Hot path: single relaxed atomic increment, no lock.
-        status_code_counts_[status_code].fetch_add (1, std::memory_order_relaxed);
+        status_code_counts_[static_cast<size_t> (status_code)].fetch_add (
+        1, std::memory_order_relaxed);
         return;
     }
     // Out-of-range (non-standard) code: dead path for real HTTP traffic.
@@ -745,7 +747,8 @@ void MetricsCollector::record_status_code (int status_code) {
 std::map<int, size_t> MetricsCollector::status_code_distribution () const {
     std::map<int, size_t> result;
     for (int code = 0; code < STATUS_CODE_SLOTS; ++code) {
-        size_t count = status_code_counts_[code].load (std::memory_order_relaxed);
+        size_t count =
+        status_code_counts_[static_cast<size_t> (code)].load (std::memory_order_relaxed);
         if (count > 0) {
             result[code] = count;
         }
